udev: beaglebone: add on-board ve-direct and mkx
For the beaglebone, half of the tty's are defined in the /etc/venus/xxxx
files and others are in udev machine.rules. Move them all to udev.

note: the mkx has a no-settings-suffix, to keep the name of the settings
file the same. builtin-mkx on newer machines do have the suffix.