udev: ccgx: add on-board ve-direct and mkx
Like the other machines, move assigning services to the on-board
peripherals to udev. The ttyO0 is special since it can also be used
as console, so add a special service to it and create a symlink so
the tty doesn't need to be hardcoded in a file. Adjust vegetty to
use the symlink.

note: the mkx has a no-settings-suffix, to keep the name of the settings
file the same. builtin-mkx on newer machines do have the suffix.