node-red: move flows-$host.json files
Under a default config, flows are named flows_<hostname>.json, which
is rather fragile, if the hostname changed the flows are lost e.g. and
hence it was changed to flows.json. When encountering flows under the
default name  and there is no flows file yet, rename them, so the flows
remain available.