node-red-contrib-victron 1.4.24 -> 1.4.25
This release features a big change in how all of the nodes are found on the underlying
dbus. Instead of looking at the fourth field of the dbus service to identyfiy
the device, it now looks a the first three fields of the dbus service path, in
combination with the the `/DeviceInstance`. This allows the device to be
re-found when the USB or VE.Direct ports change (eg at reboot). It also means
that all existing flows need to be migrated. This is implemented by a blue
'please migrate' dot next to the nodes. As long as the user has not migrated
and the path still exists on the dbus, the flows remain backwards compatiable and
thus functional.

Full changelog:
- Added input-pump and output-pump node (issue #146)
- Added "Contactor check" state for Lynx Smart BMS
- Moved systemname setting from output-system to output-settings (issue #152)
- Show system relay as "Venus System" for input-relay
- Added '/Mode' option to input-battery
- Select devices based on /DeviceInstance (issue #144)
- No longer safe all the paths in flows.json. Decreasing the size of the flows.json file.
- Set default msg.topic to be th service name + path name.
- Reselect previous path when changing service
- Rename Multi to MultiRS (issue #150)
- output-battery added and power reading on input-battery
- added example for monitoring diskspace via mqtt
- /TimeToGo is in seconds and not hours for input-system and input-battery
- Updated documentation and some minor textual fixes