dbus: add a setting for enabling tcp
In the node-red documentation it is mentioned that the dbus can be accessed
over tcp/ip. note: dbus doesn't support encryption of the tcp connection, so
this is obviously insecure. When an user makes a mistake when modifying the
config file though, the device will hardly boot, since most processes depend
on the dbus. Furthermore, the modificition will no longer be there after a
firmware update.

So instead this adds a setting to enable that inside the image. Since
localsettings is not started yet / depends on the dbus, the xml files is
read directly.