dbus-generator-starter v1.5.3
Publish the total accumulated runtime for the generator on the dbus service
(in addition to the existing location in localsettings).

This does away with another Relay/FischerPanda distinction, treating all
gensets the same. This is in anticipation of a report on VRM that will need
this data in a more unified form.