update-data: pass extraction result to the post-hook
As mentioned in [1] the post-hook is executed even if extraction the data
failed. Never running the post-hook though changes the current behaviour.
Furthermore running the post-hook might be useful for some cases, even in
case of failure, to undo something the pre-hook did. Hence this passes
either 'success' or 'extraction-failed' to the post-hook as an argument
instead, so it can act appropriately.

https://github.com/victronenergy/meta-victronenergy/commit/1f71077bdce36ab91387b2d37c3bbfdb7b4a1b11