check-updates: add support for hardware-compatibility
Whenever a new product is launched, new images are created for them,
because when adding support to an existing image would also allow
older images to be installed, which would brick the product, since
e.g. its dtb doesn't exist in the older images.

So add support for a hardware-compatibility check based on the name
in the device tree, to prevent downgrading to an unsupported version.