prepare for merging images
Support for the Cerbo GX S can be moved to the einstein image if the hardware
compatible check is added. However when testing that, candidate will switch
to using the einstein image before the release image supports the Cerbo GX S
and hence switching from candidate to release will then make the device
unbootable. So include Cerbo GX S support first, before changing the Cerbo GX S
to use the einstein image. The same holds for the ekrano, but for the nanopi
image in that case.