Add Cache-Control header to nginx to fix cache issues
The MFD app (/app) was loaded from cache some times, preventing users
from seeing new versions. This cache directive still allows browsers to
cache, but they will revalidate with a conditional request.

Confirmed in the Nginx access log that loading the MFD app now shows
entries for all files loaded, with a 304 'Not Modified' response when
the file didn't change, and a normal 200 when they did.