linux-firwmare: fix the mt7662 blob
The dunfell update included a newer version of linux firmware, in which the
mediatek files are moved to a subdirectory. The original files have become
symlinks. If the files themselves are not added to FILES:${PN}-mt7662 a runtime
dependency is added to the general linux-firware and hence all firmware. This
too large to fit. So pack the new files in a right package.