flashmq: shorter session expiration interval
The default of a month is somewhat excessive.

I'd like to encourage the use of sessions (the correct way, with random
client IDs on app start), because it makes dealing with link disconnects
easier (you don't have to hassle with re-creating subscriptions, for
instance), but then we should be more pro-active about cleaning them all
up.