add a venus-data version which restore the firmware
This will remove any rclocal files from the data partition and install a
swu file which should be included separately on the media. If the device
has a buzzer there will be a short beep to indicate it is starting. When
completed successfully there will be two sort beeps a second after each
other. When an error occured, there will be a 3 second long beep.