increase the overhead factor to 1.20
The overhead factor was reduced to make the large image fit for the Cerbo.
Unfortunately there was so little free space left that fstab could no longer
be written. Even worse awk doesn't handle this error and exits successfully,
resulting in an empty fstab. That makes the device unable to successfully
reboot. This increases the free space a bit, so at least calling the script
succeeds.