resize2fs: modify /etc/fstab after resizing
We managed to create an image with so little free space, that modifying /etc/fstab was
not possible. Since busybox its awk doesn't error about that, we ended up with an empty
fstab. This moves the modification of fstab till after the resizing and adds a
workaround for the lack of error reporting.