dbus-systemcalc-py v2.140
Show start- and end timestamp of last schedule on dbus, for logging
by VRM. On sites without DESS enabled, the paths remains invalid, so
vrmlogger will not log it. The start and end timestamp of the last
configured schedule is displayed, even if none matches at the moment.
Increase number of slots to 12.