machine-config: don't write a empty unique-id to a file
This is old / odd issue. Typically a unique-id is read from device on the SOC,
e.g. a mac address uniquely determining the device world wide. Apperently this
might fail on some rpi's during boot.

This prevents that an empty unique-id file is created, so it obvious that there
is no unique-id. It does error about it; Devices of Victron Energy shouldn't
behave like that.

https://github.com/victronenergy/venus/issues/640