dbus-systemcalc-py v2.157
Bugfixes for hysteresis bug
  - After charging the system would switch to discharge mode for 5 seconds before
    going to Idle.
  - If badly timed, system could get stuck in discharge mode if the schedule ended
    right then.

Improvements to the charge/discharge rate calculation. Charge or discharge will
never slow down, only speed up if it falls behind. This is better, as the first
calculation over a longer time period is more stable than subsequent calculations,
and the worst that can happen is we get there early.