dbus-generator-starter v1.6.11
Changes made so that dbus-generator and systemcalc don't try to control ignoring of AC-in
dbus-generator needs to ignore AC-in so that generators can warm-up and cool down.
systemcalc needs to ignore AC-in to pre-emptively disconnect the grid when there are
scheduled outages (load shedding). Therefore dbus-generator must only control AC-inputs
configured as generator, even when un-ignoring AC-in.