nginx: fix revalidation of the gui-v2 wasm version
The add_header in a child _replaces_ the add_header's off the parent instead of appending
to them. Since the no-cache header was hence missing browser would simply use the cached
version instead of validating with the host if it is not modified. This add the no-cache
header back for the gui-v2.