dbus-systemcalc-py v2.161 & v2.162
DynamicEss: Fix bug where AC-coupled PV on the output was not used to
charge the battery but fed into the grid instead. This happens when the
coping strategy is self-consumption, but there is a restriction on
charging from the grid. Bug was introduced in Venus 3.30.

Prepare for MeterBoost batteries: Force good settings for MeterBoost
batteries. Only works once newer can-bus-bms that includes detecting the
meterboost is done.

https://github.com/victronenergy/venus-private/issues/365