machine-runtime: make ve-set-passwd compatible with VRM
The python version of bcrypt always use rand() to create a salt. VRM however assumes the
salt is derived from the unique-id and hence such password hashes are not compatible with
VRM. So replace the python bcrypt version with the c version of bcrypt and pass the
unique-id as a salt. Adjust ve-set-passwd-to-pincode accordingly.