scarthgap: glib-2.0: update patches
g_file_set_contents: always make sure the file is on the disk.
The upstream g_file_set_contents didn't fully guarantee the file is
actually on the disc and for performance reasons didn't want to alter
the function, but added g_file_set_contents_full instead. Since venus
guaranteed it was on disc, lets alter g_file_set_contents to make sure
it does. As far as I remember conmann is/was the only user of it.