openssh: Add script to dynamically generate authorized keys
This is currently only for the user vnctunnel, but is not inherently
limited to that. There are conceivable other uses. All the same, it's
written defensively currently by ony working for the vnctunnel user.

The need for the vnctunnel is that we want to allow port forwarding to
IPs we don't know in advance. The 'permitopen' format is also very
limited, so any kind of netmask matching is not possible.

As noted in the script, it's not slow, despite using dbus. It takes less
than half a second to find the IP of one EV charger. And only for the
user vnctunnel.

Author: Wiebe Cazemier <wiebe@ytec.nl>