gui-v2-webassembly: add a script to determine the checksum
For normal releases a static checksum is shipped in the image and VRM uses that to
load a cached version instead of loading it from the target. During development etc
the wasm might be replaced though. This scripts allows to calculcate the same
checksum on the target itself, which can be reported to VRM.