WIP signalk-server: update npm-shrinkwrap.json
This results in many package updates. And I verified one, signalk-to-nmea2000, and
for that, unlike the check the other day with acorn, we are shipping the version
specified by npm-shrinkwrap.json. Which is good wrt that npm-shrinkwrap works for that
but not good in that we're not shipping the latest version of signalk-to-nmea2000.

Venus OS v3.40~32 has        signalk-to-nmea2000 v2.16.0
npm-shrinkwrap specified     signalk-to-nmea2000 v2.16.0 (OK)
npm-shrinkwrap now specifies signalk-to-nmea2000 v2.18.0
(and making a new build also has v2.18.0)

And meanwhile in my npm-shrinkwrap.json, the acorn entry has disappeared..

Venus OS v3.40~32 has         acorn v8.12.0
npm-shrinkwrap specified      acorn v8.10.0
npm-shrinkwrap now specifies  nothing