dbus-acsystem v0.16
Minor code cleanups, splitting into smaller modules
Improve how CurrentLimitIsAdjustable is summarised for the whole system,
in case it differs between units. In the previous version, it is possible
for this to differ between units, if they are different at startup and
depending on order of detection.

Mirror and synchronise the AC input type indication (grid, genset, shore).
It is better for this to be synchronised. Systemcalc needs this information
to present correct information to the GUI/VRM.