dbus-systemcalc-py v2.199
Fix confusing system state display, when system indicates Off even though the
Inverter/charger is on. This is caused in systems that has the ESS mode set to
External Control. In such cases, the intent was to not show any state at all,
since the state depends on what the user's control loop is doing. To prevent
this, we now rather show "External Control" as the system state. It is
ambiguous, since this state is also used when the system is controlled by a
managed battery, but it is less confusing than showing Off.

The original bug, where ESS flags caused confusion, remains fixed.
    https://github.com/victronenergy/venus/issues/1270

Closes https://github.com/victronenergy/venus-private/issues/438