checkupdates.sh: add feed and status-file arguments
This adds arguments to check a specific feed (-feed [release,..] etc)
instead of the Venus OS setting and allow to write the result to a
different file given by -status-file.

JHO: drop hardcoded assumptions about the release feed and move that
to venus-platform instead. Pass the feed as string instead of index
number.