udev: bluetooth: make sure btmgmt has a stdin
btmgmt can hang when run from udev. Commit 21a5b427de44092b38c
"udev: bluetooth: start bt-config in a shell" tried to solve that by
starting a proper shell for the script, but that prevents the script from
blocking since it is not executed in the first place, since an absolute
path needs to be given for udev to correctly run it. So instead pipe yes
into btmgmt. btmgmt doesn't actually read from stdin though, but without
it, it can hang indefinitely.

For completeness, this is introduced somewhere after bluez version v5.55,
that version works without this workaround.