swupdate-scripts: add a script to report the mmc / storage size
This script reports the size of the flash chip. Since it will report
a valid value for any block device it is booted from and to make clear
it is not a partition size, it is named storage-size.

The script will fail when the root partition is not a block device, e.g.
on the devices using ubi.