havegd: restore the init service
OE deleted the services so haveged wasn't started anymore causing booting
of the ccgx to fail. Starting haveged from udev can cause boot to hang if
there is no RTC battery, it is a mystery why that is. This adds init support
for haveged back, so booting works again (also without a battery).