watchdog: don't trigger when failing to get the timeout setting
The vrm timeout should only trigger when it is sure that is enabled.
E.g. in testmode, localsettings is not started and the script would
trigger a reboot.

Since the watchdog is running before the dbus is started, ignore the
error ouput of the dbus call, it causes unnecesarry traces during any
normal boot.