openssl: disable TLS 1, TLS 1.1 and DTS 1 system wide
See e.g. RFC 8996, these versions have been deprecated, and marked as
"Do Not Use", so lets make sure they aren't even available.

Don't recommend the legacy module, there is likely a good
reason it is marked as legacy and venus doesn't need it.

Set the default available ciphers to what mozilla recommends in
https://wiki.mozilla.org/Security/Server_Side_TLS

Make the ciphers settings commonly available, so it is possible to
use in order packages as well.

https://github.com/victronenergy/venus/issues/1406