dbus-systemcalc-py v2.213
DESS: Introduction of Reactive and split-coping strategies, allowing
the system to use the VRM schedule for its planning, but to better
take local conditions into account. For example when there is more
solar then expected, it can decide to put that exceed the scheduled
target SOC and store that excess into the battery too. (Daniel Manner
patches).
DESS: Synchronise a part of the VRM configuration with the Venus
installation, allowing the system to make information based decisions
on how to act when the forecast doesn't match reality.
DESS: Support 48 schedule slots (in preparation of 15 minute
schedules).
Add extra unit tests.