flashmq: allow a comma separated Upgrade header
This allows (older version of) Victron Connect to connect to the
mqtt-websocket again. This did no longer work because:

- Victron Connect sends a request with multiple Upgrade headers
- Newer nginx versions fold these in a single comma seperated one
- FlashMQ didn't like the comma separated Upgrade header

This patches FlashMQ to tolerate the comma separated header, so
it works again with the current Victron Connect version and a newer
nginx version. It is trivial to alter Victron Connect to not send
such requests though.