gobject-introspection: make it a bbappend
This gir / typelib stuff is rather complicated, since it has native /
cross dependencies and might require initial versions to prevent
circular dependencies. That is done in newer OE versions. This only
splits the tools in a seperate package, but does need to fiddle with
RDEPENDS to make that work.

Just make it a bbappend, since that saves us enough space and it is
no longer in our branch of oe-core.