connman: always use hwclock --noadjfile
The util-linux version of hwclock in the large image tries to write to
/etc/adjtime which will fail with a read-only rootfs. The busybox version
used in the normal image doesn't support that, nor the --noadjfile argument.
So always use the util-linux version and add --noadjfile, so it works for
both images. Also add util-linux-hwclock to the base, so this choice is
independend of the connman_%.bbappend.