npm-online-install: add update_shrinkwrap task
e.g. `bitbake -c update_shrinkwrap node-red` will update the
the shrinkwrap file in the meta itself.

Note: when using npm install and there are incompatible problems
it will update the lock file to resolve the issue. Hence switch
to clean-install, which errors if problems are encountered.

Do make sure the npm-shrinkwrap.json is available in the TMP dir,
since the the clean-install will now complain if it is not there
and for some reason it is there for node-red, but not
node-red-contrib-victron.