drop extracting to UNPACKDIR
By setting UNPACKDIR to a different value the files would be extracted
to that path. That works because venus has a patch on openembedded-core
to make that work. Upstream made more incompatible changes to the
behaviour of S etc, so it can be kept compatible anyway.

So drop support for a functional UNPACKDIR and just set it to WORKDIR,
so the patch from openembedded-core can be dropped.

Recipes which compile things needs some adjustment otherwise OE will
warn about TMPDIR in debug files.