Sets up an initial flow for node-red, if no flow exists yet
This relates to https://github.com/victronenergy/venus/issues/1506, and should
fix the issue that nodes cannot be configured on a fresh system before the
first deploy from node-red.

I tested this change locally.

This is not the full solution yet: In node-red, on deploy, I now get a warning
about an unused config node. This may have to be fixed form the
node-red-contrib-victron package, I am investigating that.

See also this discussion on the node-red forum:
https://discourse.nodered.org/t/creating-a-config-node-on-demand-import-broken/98188/11
... I tried to resolve this config node creation issue from within
node-red-contrib-victron, but so far without a feasable and clean solution.