u-boot-fw-utils_venus: Raspberry Pi
A patch is needed to sync the default U-Boot environment with the
newer U-Boot version used on Raspberry Pi 5.

Newer versions of SWUpdate use libubootenv, while older ones use
u-boot-fw-utils. This workaround allows Venus OS, which still uses
a 2016 version of SWUpdate with u-boot-fw-utils, to be built and
used with the newer U-Boot found on the Raspberry Pi 5.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>