dbus-systemcalc-py v2.217-rev6
Fix bug where charge current limit on the Multi in systems without a BMS,
or systems with a VE.Bus BMS, would not be correctly applied.

Switching the Multi off, and then adjusting the Maximum charge current limit
in the DVCC menu, would not be applied when the Multi was turned back on.
Instead the Multi would charge at the maximum it was configured for.

https://github.com/victronenergy/venus/issues/1496