Fix caching wasm file
The 'no-cache' that was there before may allow the browser to store it
in the cache, but it's not actually used to do conditional requests.

That is, it's misleading, because if the file has ever been cached
before when 'cache-control' was different, the browser will still do a
conditional request. However, on a clean browser, 'no-cache' will cause
the browser never to do a conditional request.